+++ src/main/java/com/buuz135/industrial/proxy/client/ClientProxy.java
@@ -1,6 +1,8 @@
 package com.buuz135.industrial.proxy.client;
 
+import com.buuz135.industrial.entity.EntityPinkSlime;
 import com.buuz135.industrial.proxy.CommonProxy;
+import com.buuz135.industrial.proxy.client.entity.RenderPinkSlime;
 import com.buuz135.industrial.proxy.client.event.IFTextureStichEvent;
 import com.buuz135.industrial.proxy.client.render.ContributorsCatEarsRender;
 import com.buuz135.industrial.utils.Reference;
@@ -56,12 +58,9 @@ public class ClientProxy extends CommonProxy {
 
         MinecraftForge.EVENT_BUS.register(new MobRenderInPrisonHandler());
         MinecraftForge.EVENT_BUS.register(new IFTextureStichEvent());
+
     }
 
-    @Override
-    public void postInit() {
-        super.postInit();
-    }
 
     @Override
     public void init() {
@@ -83,5 +82,13 @@ public class ClientProxy extends CommonProxy {
             e.printStackTrace();
         }
 
+        manager.entityRenderMap.put(EntityPinkSlime.class, new RenderPinkSlime(manager));
     }
+
+    @Override
+    public void postInit() {
+        super.postInit();
+
+    }
+
 }