+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 + Fixed [#19](https://github.com/Buuz135/Industrial-Foregoing/issues/19)
 + Added a Villager Trade Exchanger that trades with villagers automatically
 + Added JEI Handler to show how much power it would produce something in the Petrified Fuel Generator
++ Added Pink Slime Entity
 
 #1.2.1
 + Fixed issue [#37](https://github.com/Buuz135/Industrial-Foregoing/issues/37)